seperated the db from the sever to another file
