+++ index.js
@@ -1,12 +1,14 @@
 const Server = require('./src/Server');
 const startCronJob = require('./src/cronJob/cronJob');
 const listenQueue = require('./src/RMQ/reciverQueueMessage');
+const connectDb = require('./src/mongoConnection');
 
 const initServer = async () => {
   try {
-    await new Server().start();
+    await connectDb();
     await listenQueue();
     await startCronJob();
+    await new Server().start();
   } catch (err) {
 
   }
